requirements.txt
numpy
pandas
matplotlib
Pillow
opencv-python
torch
torchvision
torchmetrics
lightning
datasets
transformers
sentencepiece
evaluate
jiwer
python-Levenshtein
easyocr
protobuf
